% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalibrate.R
\name{ciCalibrate}
\alias{ciCalibrate}
\title{Calibrate confidence intervals to support intervals}
\usage{
ciCalibrate(
  ci = NULL,
  ciLevel = 0.95,
  estimate = mean(ci),
  se = diff(ci) * 0.5/stats::qnorm(p = 0.5 * (1 + ciLevel)),
  siLevel = 1,
  method = c("SI-normal", "SI-normal-local", "mSI-all", "mSI-normal-local",
    "mSI-eplogp"),
  priorMean,
  priorSD
)
}
\arguments{
\item{ci}{Confidence interval}

\item{ciLevel}{Confidence level}

\item{estimate}{Parameter estimate}

\item{se}{Standard error of the estiamte}

\item{siLevel}{Support level}

\item{method}{Calibration method}

\item{priorMean}{Prior mean}

\item{priorSD}{Prior standard deviation}
}
\value{
A supInt object
}
\description{
This function TODO write documentation
}
\examples{
## confidence interval of hazard ratio needs to be transformed to log-scale
ciHR <- c(0.75, 0.93)
ci <- log(ciHR)

## need prior under alternative hypothesis H1
m <- log(0.8)
s <- 2

## compute 3 support interval
si <- ciCalibrate(ci = ci, method = "SI-normal", priorMean = m, priorSD = s,
                  siLevel = 10)
si # on logHR scale
exp(si$si) # on HR scale

## plot Bayes factor function and support interval
plot(si)

## minimum support interval
msi <- ciCalibrate(ci = ci, method = "mSI-normal-local")
plot(si)

}
\author{
Samuel Pawel
}
