% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalibrate.R
\name{print.supInt}
\alias{print.supInt}
\title{Print method for supInt object}
\usage{
\method{print}{supInt}(x, ...)
}
\arguments{
\item{x}{A supInt object}

\item{...}{Other arguments}
}
\description{
Prints parameter estimate, confidence interval, and support
    interval.
}
