% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalibrate.R
\name{plot.supInt}
\alias{plot.supInt}
\title{Plot method for supInt object}
\usage{
\method{plot}{supInt}(x, xlim = x$estimate + c(-1, 1) * 3 * x$se, ...)
}
\arguments{
\item{x}{A supInt object}

\item{xlim}{Limits of x-axis}

\item{...}{Other arguments}
}
\description{
Plot method for supInt object
}
