% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciCalibrate.R
\name{print.supInt}
\alias{print.supInt}
\title{Print method for class \code{supInt}}
\usage{
\method{print}{supInt}(x, ...)
}
\arguments{
\item{x}{Object of class \code{supInt}}

\item{...}{Other arguments for print}
}
\value{
Prints text summary in the console and returns an invisible copy of
    the supInt object
}
\description{
Prints parameter estimate, confidence interval, and support
    interval.
}
\author{
Samuel Pawel
}
